Add pages module
